Function number_theory::primitive_pythag_triples

Important traits for PythagTripleIterator<F>
impl<F> Iterator for PythagTripleIterator<F> where
F: Fn(PythagTriple) -> bool, type Item = PythagTriple;

pub fn primitive_pythag_triples<F>(condition: F) -> PythagTripleIterator<F> where
    F: Fn(PythagTriple) -> bool, 

An iterator over the primitive Pythagorean triples satisfying the given condition.

The condition should be such that if a given triple passes it, then so do all smaller triples.

Examples

use number_theory::primitive_pythag_triples;

let mut triples = primitive_pythag_triples(|(a, b, c)| a + b + c <= 100);

assert_eq!(triples.next(), Some((3, 4, 5)));
assert_eq!(triples.next(), Some((8, 15, 17)));
assert_eq!(triples.next(), Some((12, 35, 37)));
assert_eq!(triples.next(), Some((20, 21, 29)));
assert_eq!(triples.next(), Some((5, 12, 13)));
assert_eq!(triples.next(), Some((7, 24, 25)));
assert_eq!(triples.next(), Some((9, 40, 41)));
assert_eq!(triples.next(), None);